backend/healthcard/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django
django-cors-headers
djangorestframework
docxcompose==1.4.0
docxtpl==0.16.8
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
PyJWT==2.8.0
python-docx==1.1.0
pytz==2024.1
requests==2.31.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
twilio==9.0.0
typing_extensions==4.10.0
urllib3
uvicorn==0.28.0
yarl==1.9.4
psycopg2-binary
django-environ
django-storages
pillow
botocore==1.23.54
boto3==1.20.54
pypandoc
pdf2image
pypdfium2